+++ CHANGELOG.md
@@ -1,11 +1,17 @@
 # Change Log
 
-## 2020-08-08
+## 2021-09-13
+
+- $.clas -> $.class
+- fastLink 从 ready 事件移到 show 事件
+  修正 show 创建的 back 无效。
+
+## 2021-08-08
 
 - 修正 addCss、removeCss bug
-	使用唯一id，避免异步加载中，样式与页面错配，或样式重复！
+  使用唯一 id，避免异步加载中，样式与页面错配，或样式重复！
 
-## 2020-08-06
+## 2021-08-06
 
 - 修正 detail 先于 master 加载 bug
   $.go 两个页面时，异步下载、加载，detail 可能比 master 先下载。